(Ported forth from the writeback version).
Fix for a data corruption that can occur with non co-ordinated IOs for
a block. The requirement for this bug to occur is for concurrent
Writes and Reads *of the same block* to be sent down. Not likely to
occur in production if we go through the pagecache (or with InnoDB
doing O_DIRECT), because buffers undergoing IO will be locked.

The bug occurs because concurrent IOs (at least one of which must be a
Write, and at least one which which must be uncached in flashcache),
can get re-ordered. The fix is to handle uncached IOs totally within
flashcache and invalidate overlapping blocks on the completion of an
uncached IO.

Although this bug is unlikely to occur in production, it is a good
idea to upgrade.

Thanks to Paul Saab for finding this bug.