In WriteThru and WriteAround modes, fall back to disk on SSD errors.
Summary: Committing an old patch to handle SSD errors in writethru and writearound
modes to fall back to disk instead of returning errors back to the user. This path
needs to be extended for writeback as well (failed reads on a !DIRTY block and failed
new writes can be similarly handled. Thanks to Arne Wiebalck for providing the impetus
for this change, reworking my initial attempt and testing the patch in production.

Test Plan:

Reviewers:

CC:

Task ID: #

Blame Rev: