Fix in del_all_pids, to the existence of a pid logic.
Summary: We had commented out the code in the del_all_pids logic
where we checked to see if the pid we are deleting is currently
alive in all releases > 2.6.31. This is because the find_task_pid*
API wasn't EXPORT_SYMBOL'ed in .33. It is in .38. So resurrecting
this back for Linux releases >= 2.6.38.
Fix brought in from facebook internal flashcache repo (kinda
important for MySQL).

Test Plan:

Reviewers:

CC:

Task ID: #

Blame Rev: