Lots of flashcache IO scaling changes.
Summary:

This commit has a number of flashcache changes aimed at improving
IO scaling.
1) Uniform distribution of blocks onto cache sets (instead of the
linear mapping before). But pin a consecutive run of disk blocks
onto a cache set (for clustering when writing). This necessitates
a flashcache ssd rev bump (2->3). Backwards compatiblity is maintained.
2) New heuristic algorithm to clean dirty+fallow blocks. When we
replace a block (read or write cache miss), we opportunistically look
to see if the least recent block on LRU is dirty and if so we clean it.
This makes more room in the cache for reads.
3) New LRU algorithm - a variant of the midpoint algorithm, using
2 LRU Queues.
Thanks to Kumar for co-authorship of some of these changes and Domas
for driving this round of flashcache scaling work.

Ported forward from facebook's internal repo.

Test Plan:

Reviewers:

CC:

Task ID: #

Blame Rev: