Switch from ioctl to dm target messages for PID blacklist/whitelist manipulation
In mainline 4.4 the ioctl() callback in the DM target has been removed
in favour of a prepare_ioctl() which selects the underlying device,
all ioctls are assume to apply to that.  In discussions with upstream on
resolving this it was suggested that the correct mechanism for this
kind of target focussed ioctl is actually DM target messages.  This
patch converts the blacklist/whitelist manipulation over to these DM
target messages.

It also adds rather primative support to the flashcache_setioctl helper to
switch to DM target messages when the existing ioctls are not supported
(ENOTTY).  This is handled by calling out to dmsetup which offers a
message command to form these requests.

I would envisage it would be possible to reduce flashcache_setioctl to a
simple shell script in the future once there is no possibility of these
tools being used with a kernel supporting the ioctl only.

This should fix the hard parts of issue #215.

fixes: #215
Signed-off-by: Andy Whitcroft <apw@ubuntu.com>